Complete refactoring of Cql2ElmVisitor to leverage CqlPreprocesorElmCommonVisitor.  More cleanup.
